+++ test/testutils/getDate.js
@@ -1,0 +1,3 @@
+const getDate = (minOffset) => new Date(new Date().getTime() + (minOffset * 1000 * 60)).toISOString();
+
+module.exports = getDate;
