+++ test/testutils/index.js
@@ -1,3 +1,4 @@
 exports.convertEventToJsonApi = require('./convertEventToJsonApi');
 exports.dbUtils = require('./dbUtils');
 exports.serverMockData = require('./serverMockData');
+exports.getDate = require('./getDate');
